Rename 'http-lobby' -> 'http-server' Also rename package 'error.report' -> 'reporting.error'
