+++ http-server/build.gradle
