+++ http-server/src/main/java/org/triplea/server/reporting/error/upload/ErrorReport.java
@@ -1,4 +1,4 @@
-package org.triplea.server.error.report.upload;
+package org.triplea.server.reporting.error.upload;
 
 import java.time.Instant;
