+++ settings.gradle
@@ -1,6 +1,6 @@
 rootProject.name='triplea'
 include 'game-core'
-include 'http-lobby'
+include 'http-server'
 include 'lobby'
 include 'lobby-db'
 include 'test-common'
